% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_clusters_SSCE.R
\name{add_clusters_SSCE}
\alias{add_clusters_SSCE}
\title{Add specificity clusters}
\usage{
add_clusters_SSCE(ssce, colPal = NULL, G = seq(1, 9), xProfiles)
}
\arguments{
\item{ssce}{The specificity SingleCellExperiment object}

\item{colPal}{A possible color palette, if wanted}

\item{G}{An integer vector specifying the numbers of mixture components (clusters)
for which the BIC is to be calculated. The default is G=1:9.}

\item{xProfiles}{The expression profiles}
}
\value{
The specificity SingleCellExperiment object with the profiles and
scores per group in metadata, plus a color per group.
}
\description{
Add specificity clusters
}
\examples{
set.seed(123)
profile <- matrix(sample(seq(1, 100, by = 0.1), size = 1000, replace = TRUE), ncol=100)
rownames(profile) <- paste0('Gene-', seq(1, nrow(profile)))
out <- initSSCE(profile)
spec <- add_clusters_SSCE(out, xProfiles = profile)
}
