% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_landscape_SSCE.R
\name{add_landscape_SSCE}
\alias{add_landscape_SSCE}
\title{Add landscape view}
\usage{
add_landscape_SSCE(ssce, addClusters = FALSE)
}
\arguments{
\item{ssce}{The specificity SingleCellExperiment object}

\item{addClusters}{A boolean variable to define if clusters should be computed}
}
\value{
The specificity SingleCellExperiment object with the UMAP coordinates
and colors saved in colData
(if addClusters is TRUE, also the membership vector saved in 'cluster')
}
\description{
Add landscape view
}
\examples{
set.seed(123)
profile <- matrix(sample(seq(1, 100, by = 0.1), size = 1000, replace = TRUE), ncol=100)
rownames(profile) <- paste0('Gene-', seq(1, nrow(profile)))
out <- initSSCE(profile)
view <- add_landscape_SSCE(out)
}
