% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_lbspec.R
\name{calculate_lbspec}
\alias{calculate_lbspec}
\title{Calculate lbspec}
\usage{
calculate_lbspec(promiscuity, high_expression = FALSE)
}
\arguments{
\item{promiscuity}{The promiscuity scores obtained from
\code{get_expression_promiscuity_levels}}

\item{high_expression}{A boolean variable to prioritize genes that have a higher
expression in specific tissues compared to the rest}
}
\value{
A lbspec score per gene
}
\description{
Calculate lbspec
}
\examples{
set.seed(123)
p <- matrix(runif(20000,0,10), ncol = 10)
colnames(p) <- paste0('Column-', seq(1, ncol(p)))
rownames(p) <- paste0('Gene-', seq(1, nrow(p)))
out <- GeneSLand:::get_lb_scores(xProfiles = p,
GeneSet = rownames(p)[sample(seq(1,nrow(p)), 20)])
GeneSLand:::calculate_lbspec(out)
}
