% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_variability_SSCE.R
\name{compute_variability_SSCE}
\alias{compute_variability_SSCE}
\title{Compute the variability inside the gene set}
\usage{
compute_variability_SSCE(ssce, add_one = TRUE)
}
\arguments{
\item{ssce}{A ssce built using gene sets}

\item{add_one}{If TRUE, to the Scores vector an initial 1 will be added, to
homogeneize the starting point of multiple genes in the computation of the
dRate.}
}
\value{
The ssce with additional colData columns. \code{IntraVariability} is the
variability across bins. \code{MeanIntraVariability} is the mean of
\code{IntraVariability} when there is a change (so that the line is not 0).
\code{ScoresVariability} is a list with the sd of the lines of the genes in the
gene set.
}
\description{
Compute the variability inside the gene set
}
\examples{
#' set.seed(123)
profile <- matrix(sample(seq(1, 100, by = 0.1),
size = 10000000, replace = TRUE), ncol=10)
rownames(profile) <- paste0('Gene-', seq(1, nrow(profile)))
colnames(profile) <- paste0('Column-', seq(1, ncol(profile)))
out <- createSSCE(profile,
geneGroups = list('A' = rownames(profile)[seq(1,50)],
'B' = rownames(profile)[seq(51,100)],
'C' = rownames(profile)[seq(101,150)],
'D' = rownames(profile)[seq(151,200)],
'E' = rownames(profile)[seq(201,250)],
'F' = rownames(profile)[seq(251,300)],
'G' = rownames(profile)[seq(301,350)],
'H' = rownames(profile)[seq(351,400)],
'I' = rownames(profile)[seq(401,450)],
'L' = rownames(profile)[seq(451,500)],
'M' = rownames(profile)[seq(501,550)]))
compute_variability_SSCE(out)
}
