% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeLB.R
\name{computeLB}
\alias{computeLB}
\title{Compute the Level-Breadth scores}
\usage{
computeLB(xProfiles, GeneSet, n_bins = 50)
}
\arguments{
\item{xProfiles}{The expression profiles}

\item{GeneSet}{The gene or gene set of interest}

\item{n_bins}{Number of expression level bins}
}
\value{
The L-B scores as a vector of B scores at increasing values of L bins
}
\description{
Compute the Level-Breadth scores
}
\examples{
set.seed(123)
profile <- matrix(sample(seq(1, 100, by = 0.1), size = 1000, replace = TRUE), ncol=100)
rownames(profile) <- paste0('Gene-', seq(1, nrow(profile)))
GeneSLand:::computeLB(xProfiles = profile, GeneSet = 'Gene-1')
}
