% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSSCE.R
\name{createSSCE}
\alias{createSSCE}
\title{Create the Specificity SingleCellExperiment object}
\usage{
createSSCE(
  xProfiles,
  geneGroups = NULL,
  addClusters = FALSE,
  compute_significance = FALSE,
  compute_variability = FALSE,
  nRand = 100,
  add_one = FALSE,
  n_bins = 50
)
}
\arguments{
\item{xProfiles}{The expression profiles}

\item{geneGroups}{The gene groups to group the profiles}

\item{addClusters}{A boolean variable to define if clusters should be
computed}

\item{compute_significance}{A boolean to define if the significance of the
comparison with random should be computed}

\item{compute_variability}{A boolean to define if the intra-gene group
variability has to be computed}

\item{nRand}{The number of times the randomization is run (needed when using
geneGroups)}

\item{add_one}{If TRUE, to the Scores vector an initial 1 will be added, to
homogeneize the starting point of multiple genes in the computation of the
dRate.}

\item{n_bins}{The number of wanted bins. Defaults to 50.}
}
\value{
The Specificity SingleCellExperiment object, with AUC, dRate, lbSpec
and initB values for each gene, containing as counts the Breadth values for
the 50 bins of expression level. It will also contain the x and y coordinates
for the UMAP plot, as well as the colors. If addClusters is TRUE, also the
clusters. If compute_significance is TRUE and the SSCE is built using gene
groups, also the variability intra gene group is computed. If
compute_significance and the SSCE is built using gene groups, also the
significance against random is computed.
}
\description{
Create the Specificity SingleCellExperiment object
}
\examples{
set.seed(123)
profile <- matrix(sample(seq(1, 100, by = 0.1), size = 1000, replace = TRUE), ncol=100)
rownames(profile) <- paste0('Gene-', seq(1, nrow(profile)))
out <- createSSCE(profile)
}
