% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_drate_max.R
\name{estimate_drate_max}
\alias{estimate_drate_max}
\title{Estimate the maximum of the drate}
\usage{
estimate_drate_max(y, x = NULL)
}
\arguments{
\item{y}{The y values}

\item{x}{The x values (if NULL, it will be from 1 to length of y)}
}
\value{
The value of the maximum drate
}
\description{
Estimate the maximum of the drate
}
\examples{
GeneSLand:::estimate_drate_max(runif(100, 2, 4))
}
