% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_zero_rows.R
\name{filter_zero_rows}
\alias{filter_zero_rows}
\title{Filter zero rows}
\usage{
filter_zero_rows(m)
}
\arguments{
\item{m}{A matrix}
}
\value{
The matrix without rows that sum to 0
}
\description{
Filter zero rows
}
\examples{
set.seed(123)
profile <- matrix(sample(seq(1, 100, by = 0.1), size = 1000, replace = TRUE), ncol=100)
rownames(profile) <- paste0('Gene-', seq(1, nrow(profile)))
sub_profile <- GeneSLand:::filter_zero_rows(profile)
}
