% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_auc_SSCE.R
\name{get_auc_SSCE}
\alias{get_auc_SSCE}
\title{Calculate AUC for a ssce object}
\usage{
get_auc_SSCE(ssce, genes = NULL)
}
\arguments{
\item{ssce}{The specificity single cell experiment object}

\item{genes}{If some genes are specified, the AUC of only those genes
will be computed}
}
\value{
The AUC scores as a vector.
}
\description{
Calculate AUC for a ssce object
}
\examples{
set.seed(123)
profile <- matrix(sample(seq(1, 100, by = 0.1), size = 1000, replace = TRUE), ncol=100)
rownames(profile) <- paste0('Gene-', seq(1, nrow(profile)))
out <- createSSCE(profile)
get_auc_SSCE(out)
}
