% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_drate_SSCE.R
\name{get_drate_SSCE}
\alias{get_drate_SSCE}
\title{Calculate dRate for a ssce object}
\usage{
get_drate_SSCE(ssce, genes = NULL, add_one = TRUE)
}
\arguments{
\item{ssce}{The specificity single cell experiment object}

\item{genes}{If some genes are specified, the dRate of only those genes
will be computed}

\item{add_one}{If TRUE, to the Scores vector an initial 1 will be added, to
homogeneize the starting point of multiple genes in the computation of the
dRate.}
}
\value{
The dRate scores as a vector.
}
\description{
Calculate dRate for a ssce object
}
\examples{
set.seed(123)
profile <- matrix(sample(seq(1, 100, by = 0.1), size = 1000, replace = TRUE), ncol=100)
rownames(profile) <- paste0('Gene-', seq(1, nrow(profile)))
out <- createSSCE(profile)
get_drate_SSCE(out)
}
