% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lbspec_SSCE.R
\name{get_lbspec_SSCE}
\alias{get_lbspec_SSCE}
\title{Calculate lbSpec for a ssce object}
\usage{
get_lbspec_SSCE(ssce, genes = NULL, high_expression = FALSE)
}
\arguments{
\item{ssce}{The specificity single cell experiment object}

\item{genes}{If some genes are specified, the lbspec of only those genes
will be computed}

\item{high_expression}{A boolean that defines if high expression is favored
in the computation of lbSpec. Defaults to FALSE.}
}
\value{
The lbSpec scores as a vector.
}
\description{
Calculate lbSpec for a ssce object
}
\examples{
set.seed(123)
profile <- matrix(sample(seq(1, 100, by = 0.1), size = 1000, replace = TRUE), ncol=100)
rownames(profile) <- paste0('Gene-', seq(1, nrow(profile)))
out <- createSSCE(profile)
get_lbspec_SSCE(out)
}
