% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpa_celltype_ssce.R
\name{hpa_celltype_ssce}
\alias{hpa_celltype_ssce}
\title{Specificity Single Cell Experiment object}
\format{
A SSCE with 50 rows and 18782 columns:
\describe{
\item{assays}{A list containing one matrix named "counts" (expression bins × genes)}
\item{colData}{A \code{DataFrame} with gene-level information (scores, coordinates)}
}
}
\source{
Processed from HPA celltypes
}
\usage{
hpa_celltype_ssce()
}
\description{
It contains a Specificity Single Cell Experiment object built using Human
Protein Atlas cell types
}
