% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpa_tissue_ssce.R
\name{hpa_tissue_ssce}
\alias{hpa_tissue_ssce}
\title{Specificity Single Cell Experiment object}
\format{
A SSCE with 50 rows and 18848 columns:
\describe{
\item{assays}{A list containing one matrix named "counts" (expression bins × genes)}
\item{colData}{A \code{DataFrame} with gene-level information (scores, coordinates)}
}
}
\source{
Processed from HPA tissues
}
\usage{
hpa_tissue_ssce()
}
\description{
It contains a Specificity Single Cell Experiment object built using Human
Protein Atlas tissues
}
