% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ma_exn_marmoset_ssce.R
\docType{data}
\name{ma_exn_marmoset_ssce}
\alias{ma_exn_marmoset_ssce}
\title{Specificity Single Cell Experiment object}
\format{
A SSCE with 50 rows and 15563 columns:
\describe{
\item{assays}{A list containing one matrix named "counts" (expression bins × genes)}
\item{colData}{A \code{DataFrame} with gene-level information (scores, coordinates)}
}
}
\source{
Processed from Ma et al. marmoset excitatory neurons subtypes pseudobulk
}
\usage{
ma_exn_marmoset_ssce
}
\description{
It contains a Specificity Single Cell Experiment object built using marmoset
PFC excitatory neurons subtypes from Ma et al. (DOI: 10.1126/science.abo7257).
Only subtypes in common with the other primates species (human, chimpanzee,
rhesus) were considered.
}
\keyword{datasets}
