% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ma_glia_chimpanzee_ssce.R
\docType{data}
\name{ma_glia_chimpanzee_ssce}
\alias{ma_glia_chimpanzee_ssce}
\title{Specificity Single Cell Experiment object}
\format{
A SSCE with 50 rows and 15702 columns:
\describe{
\item{assays}{A list containing one matrix named "counts" (expression bins × genes)}
\item{colData}{A \code{DataFrame} with gene-level information (scores, coordinates)}
}
}
\source{
Processed from Ma et al. chimpanzee glia subtypes pseudobulk
}
\usage{
ma_glia_chimpanzee_ssce
}
\description{
It contains a Specificity Single Cell Experiment object built using chimpanzee
PFC glia subtypes from Ma et al. (DOI: 10.1126/science.abo7257).
Only subtypes in common with the other primates species (human, rhesus,
marmoset) were considered.
}
\keyword{datasets}
