% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ma_neu_marmoset_ssce.R
\docType{data}
\name{ma_neu_marmoset_ssce}
\alias{ma_neu_marmoset_ssce}
\title{Specificity Single Cell Experiment object}
\format{
A SSCE with 50 rows and 15648 columns:
\describe{
\item{assays}{A list containing one matrix named "counts" (expression bins × genes)}
\item{colData}{A \code{DataFrame} with gene-level information (scores, coordinates)}
}
}
\source{
Processed from Ma et al. marmoset neuron subtypes pseudobulk
}
\usage{
ma_neu_marmoset_ssce
}
\description{
It contains a Specificity Single Cell Experiment object built using marmoset
PFC neuronal subtypes from Ma et al. (DOI: 10.1126/science.abo7257). Only
subtypes in common with the other primates species (human, chimpanzee,
rhesus) were considered.
}
\keyword{datasets}
