% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ma_subtype_neuro_ssce.R
\docType{data}
\name{ma_subtype_neuro_ssce}
\alias{ma_subtype_neuro_ssce}
\title{Specificity Single Cell Experiment object}
\format{
A SSCE with 50 rows and 18782 columns:
\describe{
\item{assays}{A list containing one matrix named "counts" (expression bins × genes)}
\item{colData}{A \code{DataFrame} with gene-level information (scores, coordinates)}
}
}
\source{
Processed from Ma et al. human neuron subtypes pseudobulk
}
\usage{
ma_subtype_neuro_ssce
}
\description{
It contains a Specificity Single Cell Experiment object built using human
PFC neuronal subtypes from Ma et al. (DOI: 10.1126/science.abo7257)
}
\keyword{datasets}
