% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lines_SSCE.R
\name{plot_lines_SSCE}
\alias{plot_lines_SSCE}
\title{Plot the LB lines given a ssce}
\usage{
plot_lines_SSCE(
  ssce,
  gene,
  addLines = FALSE,
  Rand = TRUE,
  Intra = TRUE,
  where_zero = FALSE,
  gene_color = "black",
  linewidth = 2,
  ...
)
}
\arguments{
\item{ssce}{The Specificity SingleCellExperiment object.}

\item{gene}{The gene or gene set to plot (a column in the SSCE)}

\item{addLines}{If TRUE, the x and y mid points are shown as lines}

\item{Rand}{If TRUE, the random distribution mean (+ and - the SD) are shown.
Valid only if the SSCE is built using gene sets.}

\item{Intra}{If TRUE, the intra gene set distribution mean + and - the SD are
shown. Valid only if the SSCE is built using gene sets.}

\item{where_zero}{If TRUE, a vertical line to show where the line arrives at
zero is plotted.}

\item{gene_color}{The color of the main line}

\item{linewidth}{The line width of the main line (\code{lwd} in plot)}

\item{...}{Additional graphical parameters.}
}
\value{
A plot of the LB line
}
\description{
Plot the LB lines given a ssce
}
\examples{
set.seed(123)
profile <- matrix(sample(seq(1, 100, by = 0.1), size = 1000, replace = TRUE), ncol=100)
rownames(profile) <- paste0('Gene-', seq(1, nrow(profile)))
out <- createSSCE(profile)
plot_lines_SSCE(out, 'Gene-1')
}
