% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_promiscuity_lines.R
\name{plot_promiscuity_lines}
\alias{plot_promiscuity_lines}
\title{plot_promiscuity_lines
A function to plot the promiscuity scores obtained using
\code{get_expression_promiscuity_levels}}
\usage{
plot_promiscuity_lines(prom_out, addLines = FALSE, Rand = TRUE, ...)
}
\arguments{
\item{prom_out}{The resulting object of \code{get_expression_promiscuity_levels}}

\item{addLines}{A boolean variable to define if the lines dividing the plot
in quadrants have to be plotted or not}

\item{Rand}{A boolean variable to define if the random scores lines (mean and
stats::sd) have to be plotted}

\item{...}{The plot standard parameters}
}
\value{
The plot with promiscuity lines
}
\description{
plot_promiscuity_lines
A function to plot the promiscuity scores obtained using
\code{get_expression_promiscuity_levels}
}
\examples{
set.seed(123)
p <- matrix(runif(20000,0,10), ncol = 10)
colnames(p) <- paste0('Column-', seq(1, ncol(p)))
rownames(p) <- paste0('Gene-', seq(1, nrow(p)))
out <- GeneSLand:::get_lb_scores(xProfiles = p,
GeneSet = rownames(p)[sample(seq(1,nrow(p)), 20)])
GeneSLand:::plot_promiscuity_lines(out)
}
