% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSSCE.R
\name{plotSSCE}
\alias{plotSSCE}
\title{Plot the Specificity SingleCellExperiment object}
\usage{
plotSSCE(ssce, values = NULL, size = 3, genes = NULL, ...)
}
\arguments{
\item{ssce}{The Specificity SingleCellExperiment object}

\item{values}{The names of the colData parameter to show or vector of values
to plot. It has to be numeric. If NULL, the standard UMAP will be plotted.}

\item{size}{The size of the points. Defaults to 3.}

\item{genes}{Which, if any (NULL), genes should be displayed in the UMAP.}

\item{...}{Additional graphical parameters for geom_label_repel}
}
\value{
The ggplot of the SSCE
}
\description{
Plot the Specificity SingleCellExperiment object
}
\examples{
set.seed(123)
profile <- matrix(sample(seq(1, 100, by = 0.1), size = 1000, replace = TRUE), ncol=100)
rownames(profile) <- paste0('Gene-', seq(1, nrow(profile)))
out <- createSSCE(profile)
plotSSCE(out)
}
