% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_scores.R
\name{random_scores}
\alias{random_scores}
\title{Get random scores}
\usage{
random_scores(xProfiles, geneGroups, nRand = 100, n_bins = 50)
}
\arguments{
\item{xProfiles}{The profiles on which to build the SSCE}

\item{geneGroups}{The geneGroups used to build the SSCE}

\item{nRand}{The number of randomizations to use}

\item{n_bins}{The number of expression level bins used}
}
\value{
A list with the random scores for each unique length in the
geneGroups list, named using 'N' plus the length.
}
\description{
Get random scores
}
\examples{
set.seed(123)
profile <- matrix(sample(seq(1, 100, by = 0.1), size = 10000000, replace = TRUE), ncol=10)
rownames(profile) <- paste0('Gene-', seq(1, nrow(profile)))
colnames(profile) <- paste0('Column-', seq(1, ncol(profile)))
out <- random_scores(profile,
geneGroups = list('A' = rownames(profile)[seq(1,50)],
'B' = rownames(profile)[seq(51,100)],
'C' = rownames(profile)[seq(101,150)],
'D' = rownames(profile)[seq(151,200)],
'E' = rownames(profile)[seq(201,250)],
'F' = rownames(profile)[seq(251,300)],
'G' = rownames(profile)[seq(301,350)],
'H' = rownames(profile)[seq(351,400)],
'I' = rownames(profile)[seq(401,450)],
'L' = rownames(profile)[seq(451,500)],
'M' = rownames(profile)[seq(501,550)]))
}
