% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitSSCE.R
\name{splitSSCE}
\alias{splitSSCE}
\title{Split the Specificity SingleCellExperiment object based on a membership vector}
\usage{
splitSSCE(ssce, colVec)
}
\arguments{
\item{ssce}{The specificity SingleCellExperiment object}

\item{colVec}{The membership vector}
}
\value{
The splitted ssce
}
\description{
Split the Specificity SingleCellExperiment object based on a membership vector
}
\examples{
set.seed(123)
profile <- matrix(sample(seq(1, 100, by = 0.1), size = 1000, replace = TRUE), ncol=100)
rownames(profile) <- paste0('Gene-', seq(1, nrow(profile)))
out <- initSSCE(profile)
splitSSCE <- splitSSCE(out, colVec = c(rep('A', 3), rep('B', 2), rep('C', 5)))
}
