% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_gene_lbspec.R
\name{calculate_gene_lbspec}
\alias{calculate_gene_lbspec}
\title{Calculate lbspec for a single gene}
\usage{
calculate_gene_lbspec(scores, high_expression = FALSE)
}
\arguments{
\item{scores}{The gene expression breadth}

\item{high_expression}{A boolean variable to prioritize genes that have a higher
expression in specific tissues compared to the rest}
}
\value{
A lbspec score
}
\description{
Calculate lbspec for a single gene
}
\examples{
set.seed(123)
p <- matrix(runif(20000,0,10), ncol = 10)
colnames(p) <- paste0('Column-', seq(1, ncol(p)))
rownames(p) <- paste0('Gene-', seq(1, nrow(p)))
P <- as.matrix(p)
Pc <- as.matrix(P[intersect(rownames(p),rownames(P)),])
Levs <- seq(0, max(P), length.out=50)
Scores <- sapply(Levs, function(i) {mean(rowMeans(Pc>i))})
calculate_gene_lbspec(Scores)
}
