% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_expression_specificity.R
\name{compute_expression_specificity}
\alias{compute_expression_specificity}
\title{Get a specificity score per column of \code{xProfiles} per gene}
\usage{
compute_expression_specificity(xProfiles, ssce)
}
\arguments{
\item{xProfiles}{The expression profiles}

\item{ssce}{A Specificity Single Cell Experiment object}
}
\value{
A specificity score matrix
}
\description{
Get a specificity score per column of \code{xProfiles} per gene
}
\examples{
set.seed(123)
profile <- matrix(sample(seq(1, 100, by = 0.1), size = 1000, replace = TRUE), ncol=100)
rownames(profile) <- paste0('Gene-', seq(1, nrow(profile)))
out <- createSSCE(profile)
compute_expression_specificity(profile, out)
}
