% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_drate_mean.R
\name{estimate_drate_mean}
\alias{estimate_drate_mean}
\title{Estimate the mean drate}
\usage{
estimate_drate_mean(y, x = NULL, add_one = FALSE)
}
\arguments{
\item{y}{The y values}

\item{x}{The x values (if NULL, it will be from 1 to length of y)}

\item{add_one}{If TRUE, to the y vector an initial 1 will be added, to
homogeneize the starting point of multiple ys.}
}
\value{
The mean of the drate
}
\description{
Estimate the mean drate
}
\examples{
estimate_drate_mean(runif(100, 2, 4))
}
