% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lb_scores.R
\name{get_lb_scores}
\alias{get_lb_scores}
\title{get_lb_lines_scores
Defines the promiscuity levels of a gene or a gene set \code{GeneSet} in a given
set of profiles \code{xProfiles}, defined the number of expression values to use
as a threshold \code{n_bins}}
\usage{
get_lb_scores(
  xProfiles,
  GeneSet,
  nRand = 100,
  n_bins = 50,
  estRand = TRUE,
  highExpression = FALSE,
  random_list = NULL,
  add_one = TRUE
)
}
\arguments{
\item{xProfiles}{The set of profiles to be used}

\item{GeneSet}{The gene or gene set of interest}

\item{nRand}{Number of random iterations (defaults to 100)}

\item{n_bins}{The number of expression values to use as threshold to
compute the promiscuity (defaults to 50)}

\item{estRand}{A boolean variable to determine if the random baseline has to
be computed}

\item{highExpression}{A boolean to define whether the lbSpec needs to
be weighted on high expression or not}

\item{random_list}{If the GeneSet is of length > 1, you can provide already
a list of score matrices obtained by randomization. Each element of the list
has to be obtained by using the length of the GeneSet (or GeneSets) and be
named using 'N' and the length of the GeneSet (if list of length 1 or
GeneSets otherwise).}

\item{add_one}{If TRUE, to the Scores vector an initial 1 will be added, to
homogeneize the starting point of multiple genes in the computation of the
dRate.}
}
\value{
A list containing the promiscuity scores (\code{scores}),
the decreasing rate (\code{dRate}) and the specificity (\code{lbSpec}). If a gene set
is given, also the scores for each gene in the gene set are given.
If estRand=T, also the random promiscuity scores (\code{rscores}), the z-score
(\code{z}) and p-values (\code{p}) of the difference with random and the random scores
of each iteration (\code{random_list}).
}
\description{
get_lb_lines_scores
Defines the promiscuity levels of a gene or a gene set \code{GeneSet} in a given
set of profiles \code{xProfiles}, defined the number of expression values to use
as a threshold \code{n_bins}
}
\examples{
set.seed(123)
p <- matrix(runif(20000,0,10), ncol = 10)
colnames(p) <- paste0('Column-', seq(1, ncol(p)))
rownames(p) <- paste0('Gene-', seq(1, nrow(p)))
out <- GeneSLand:::get_lb_scores(xProfiles = p,
GeneSet = rownames(p)[sample(seq(1,nrow(p)), 20)])
}
