% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initSSCE.R
\name{initSSCE}
\alias{initSSCE}
\title{Initialize the Specificity SingleCellExperiment object}
\usage{
initSSCE(
  xProfiles,
  geneGroups = NULL,
  nRand = 100,
  add_one = TRUE,
  n_bins = 50
)
}
\arguments{
\item{xProfiles}{The starting profiles}

\item{geneGroups}{The gene groups to group the profiles}

\item{nRand}{The number of times the randomization is run (needed when using
geneGroups)}

\item{add_one}{If TRUE, to the Scores vector an initial 1 will be added, to
homogeneize the starting point of multiple genes in the computation of the
dRate.}

\item{n_bins}{The number of wanted bins. Defaults to 50.}
}
\value{
The Specificity SingleCellExperiment object, with AUC, dRate, lbSpec
and initB values for each gene, containing as counts the Breadth values for
the 50 bins of expression level. If the SSCE is built using geneGroups, it
will also return the random mean score in rscores, the random scores list in
random_list and the scores for each gene in the gene set in intrascores.
}
\description{
Initialize the Specificity SingleCellExperiment object
}
\examples{
set.seed(123)
profile <- matrix(sample(seq(1, 100, by = 0.1), size = 1000, replace = TRUE), ncol=100)
rownames(profile) <- paste0('Gene-', seq(1, nrow(profile)))
out <- initSSCE(profile)
}
