% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapSSCE.R
\name{mapSSCE}
\alias{mapSSCE}
\title{Title}
\usage{
mapSSCE(
  new_profiles,
  ssce,
  genes,
  size = 3,
  values = "AUC",
  gene_name = "MappedGene",
  size_converter = 1.5,
  add_one = FALSE,
  ...
)
}
\arguments{
\item{new_profiles}{The whole expression matrix of the genes that you want to
map (including the genes to be mapped, but not exclusively).}

\item{ssce}{The ssce on which to map}

\item{genes}{The genes to map. It can either be a vector, and then each
single gene in the vector will be mapped independently, or a named list,
associating genes to a group, in which case the mapped points will be the
groups.}

\item{size}{Size of the dots in the plot}

\item{values}{Which score to show in the plot as a color gradient}

\item{gene_name}{If a single gene is mapped, its name can be defined here, as
the new_profiles will likely be a vector.}

\item{size_converter}{How many times bigger the mapped points are with
respect to the baseline size}

\item{add_one}{If TRUE, to the Scores vector an initial 1 will be added, to
homogeneize the starting point of multiple genes in the computation of the
dRate.}

\item{...}{Additional graphical parameters for geom_label_repel}
}
\value{
A list containing the result of \code{get_expression_promiscuity_levels}
on the new gene(group)s, so their Breadth values, the AUC, dRate and lbSpec
scores, plus the plot with the mapped points and the coordinates dataframe.
}
\description{
Title
}
\examples{
set.seed(123)
profile <- matrix(sample(seq(1, 100, by = 0.1), size = 1000, replace = TRUE), ncol=100)
rownames(profile) <- paste0('Gene-', seq(1, nrow(profile)))
out <- createSSCE(profile)
new_profile <- matrix(sample(seq(1, 100, by = 0.1), size = 1000, replace = TRUE), ncol=100)
rownames(new_profile) <- paste0('Gene-', seq(1, nrow(new_profile)))
mapped_out <- mapSSCE(new_profile, ssce = out, genes = 'Gene-1')
}
