% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscore_to_pval.R
\name{zscore_to_pval}
\alias{zscore_to_pval}
\title{zscore_to_pval
Converts Z-scores to P-values}
\usage{
zscore_to_pval(z)
}
\arguments{
\item{z}{Z-score}
}
\value{
The p-value
}
\description{
zscore_to_pval
Converts Z-scores to P-values
}
\examples{
zscore_to_pval(1.96)
}
